[commn/daavalues] add data_value_kernel from data_value
